AND-1723 Refactoring
