+++ app/src/main/java/com/tangem/tap/common/redux/global/GlobalState.kt
@@ -3,12 +3,12 @@ package com.tangem.tap.common.redux.global
 import com.tangem.domain.common.ScanResponse
 import com.tangem.tap.common.analytics.AnalyticsHandler
 import com.tangem.tap.common.entities.FiatCurrency
+import com.tangem.tap.common.feedback.FeedbackManager
 import com.tangem.tap.common.redux.StateDialog
 import com.tangem.tap.domain.PayIdManager
 import com.tangem.tap.domain.TapWalletManager
 import com.tangem.tap.domain.configurable.config.ConfigManager
 import com.tangem.tap.domain.configurable.warningMessage.WarningMessagesManager
-import com.tangem.tap.features.feedback.FeedbackManager
 import com.tangem.tap.features.onboarding.OnboardingManager
 import com.tangem.tap.network.exchangeServices.CurrencyExchangeManager
 import org.rekotlin.StateType
