+++ app/src/main/java/com/tangem/tap/features/details/ui/DetailsFragment.kt
@@ -9,6 +9,7 @@ import androidx.transition.TransitionInflater
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.tangem.domain.common.getTwinCardIdForUser
 import com.tangem.tap.common.extensions.show
+import com.tangem.tap.common.feedback.FeedbackEmail
 import com.tangem.tap.common.redux.global.GlobalAction
 import com.tangem.tap.common.redux.navigation.AppScreen
 import com.tangem.tap.common.redux.navigation.NavigationAction
@@ -16,7 +17,6 @@ import com.tangem.tap.domain.extensions.isMultiwalletAllowed
 import com.tangem.tap.features.details.redux.DetailsAction
 import com.tangem.tap.features.details.redux.DetailsState
 import com.tangem.tap.features.details.redux.SecurityOption
-import com.tangem.tap.features.feedback.FeedbackEmail
 import com.tangem.tap.features.wallet.redux.WalletAction
 import com.tangem.tap.store
 import com.tangem.wallet.R
