+++ app/src/main/java/com/tangem/tap/features/send/ui/dialogs/SendTransactionFailsDialog.kt
@@ -2,8 +2,8 @@ package com.tangem.tap.features.send.ui.dialogs
 
 import android.content.Context
 import androidx.appcompat.app.AlertDialog
+import com.tangem.tap.common.feedback.SendTransactionFailedEmail
 import com.tangem.tap.common.redux.global.GlobalAction
-import com.tangem.tap.features.feedback.SendTransactionFailedEmail
 import com.tangem.tap.features.send.redux.SendAction
 import com.tangem.tap.store
 import com.tangem.wallet.R
