+++ app/src/main/java/com/tangem/tap/features/wallet/ui/dialogs/ScanFailsDialog.kt
@@ -3,8 +3,8 @@ package com.tangem.tap.features.wallet.ui.dialogs
 import android.content.Context
 import androidx.appcompat.app.AlertDialog
 import com.tangem.tap.common.extensions.dispatchDialogHide
+import com.tangem.tap.common.feedback.ScanFailsEmail
 import com.tangem.tap.common.redux.global.GlobalAction
-import com.tangem.tap.features.feedback.ScanFailsEmail
 import com.tangem.tap.store
 import com.tangem.wallet.R
